split the json generators into separate files, fix some issues
