+++ openapi-codegen/sbt-plugin/src/sbt-test/sbt-openapi-codegen/oneOf-json-roundtrip_jsoniter/project/build.properties
@@ -1,0 +1,1 @@
+sbt.version=1.10.11
